osm.CreateEmptyRaster

Create an empty raster for an area. It downloads the area from OSM services, and then creates an empty raster of this area.

osm.CreateEmptyRaster(inPerioxi = "Mytilene Municipal Unit")

Arguments

inPerioxi

The name of an area (string) for searching the OSM services.

Value

An empty raster that covers the full area

See also

Author

Dimitris Kavroudakis dimitris123@gmail.com

Examples

library(devtools)
#> Loading required package: usethis
install_github("dimitrisk/goal")
#> Downloading GitHub repo dimitrisk/goal@HEAD
#> 
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpkJjcMk/remotes513331b9df40/dimitrisk-goal-f3dc3f2/DESCRIPTION’ ... OK
#> * preparing ‘goal’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> * building ‘goal_1.0.tar.gz’
#> 
#> Installing package into ‘/tmp/RtmpbJwbLb/temp_libpath5f1f6d613907’
#> (as ‘lib’ is unspecified)
library(goal)
result = goal::osm.CreateEmptyRaster(inPerioxi="Mytilene Municipal Unit")
#> Please note that rgdal will be retired during October 2023,
#> plan transition to sf/stars/terra functions using GDAL and PROJ
#> at your earliest convenience.
#> See https://r-spatial.org/r/2023/05/15/evolution4.html and https://github.com/r-spatial/evolution
#> rgdal: version: 1.6-7, (SVN revision 1203)
#> Geospatial Data Abstraction Library extensions to R successfully loaded
#> Loaded GDAL runtime: GDAL 3.0.4, released 2020/01/28
#> Path to GDAL shared files: /usr/share/gdal
#> GDAL binary built with GEOS: TRUE 
#> Loaded PROJ runtime: Rel. 7.0.0, March 1st, 2020, [PJ_VERSION: 700]
#> Path to PROJ shared files: /home/dimitrisk/.local/share/proj:/usr/share/proj
#> PROJ CDN enabled: FALSE
#> Linking to sp version:1.6-1
#> To mute warnings of possible GDAL/OSR exportToProj4() degradation,
#> use options("rgdal_show_exportToProj4_warnings"="none") before loading sp or rgdal.
result
#> class      : RasterLayer 
#> dimensions : 180, 360, 64800  (nrow, ncol, ncell)
#> resolution : 0.0004453886, 0.0009165467  (x, y)
#> extent     : 26.45563, 26.61597, 39.00806, 39.17304  (xmin, xmax, ymin, ymax)
#> crs        : +proj=longlat +datum=WGS84 +no_defs 
#>