osm.getFrequency

Get Frequency Table

osm.getFrequency(indf, inword = "shop", removeNA = T)

Arguments

indf

The vec

inword

Boole

removeNA

Ad

Value

A nice histogram of a single variable with an additional Cumulative Density curve.

See also

Author

Dimitris Kavroudakis dimitris123@gmail.com

Examples

library(goal)
sh = osm.getPOI(inPerioxi = "Mytilene Municipal Unit", inkey = "shop")
#> Issuing query to Overpass API ...
#> Announced endpoint: none
#> Query complete!
#> converting OSM data to sf format
shops = osm.combineShops(sh)
#> Warning: st_centroid assumes attributes are constant over geometries of x
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019
#> Warning: GDAL Error 1: PROJ: proj_as_wkt: DatumEnsemble can only be exported to WKT2:2019
#> Warning: point from first coordinate only
freq2 = osm.getFrequency(shops, inword = "shop", removeNA = F)
freq2
#> # A tibble: 85 × 3
#>    shop            n  freq
#>  * <chr>       <int> <dbl>
#>  1 NA            438 39.7 
#>  2 clothes        92  8.33
#>  3 convenience    59  5.34
#>  4 hairdresser    38  3.44
#>  5 bakery         27  2.45
#>  6 jewelry        27  2.45
#>  7 supermarket    25  2.26
#>  8 kiosk          23  2.08
#>  9 shoes          22  1.99
#> 10 butcher        21  1.90
#> # … with 75 more rows